requirements.txt
joblib==1.1.0
numpy
opencv-python==4.5.4.60
Pillow==5.1.0
pupil-detectors
pye3d
PyQt5==5.13.0
PyYAML==6.0
QDarkStyle==3.0.3
scikit-learn==0.24.2
scipy==1.5.4
screeninfo==0.8
msvc
mmdet